==> 01 - Limpeza de um semestre.ipynb <==
# Limpando os dados de um semestre
## Objetivo
Esse notebook tem como objetivo realizar a limpeza de um semestre específico, para assim criar uma função
que possa ser utilizada para outros semestres.

import pandas as pd
import numpy as np

matr = pd.read_csv("data/matricula-de-20162.csv", sep=";")

matr.head()

matr.shape

## Pegando os dados das turmas de C&T
Id do curso do Ciências e Tecnologia = 10320810

id_cet = 10320810
matr_cet = matr[matr.id_curso == id_cet]
matr_cet.shape

## Pegando apenas as matriculas realizadas

matr_cet.descricao.unique()

indesejada = ['DESISTENCIA', 'INDEFERIDO', 'EXCLUIDA', 'CANCELADO']

matriculas_cet = matr_cet[~matr_cet.descricao.isin(indesejada)]

matriculas_cet.shape

# Selecionando colunas desejadas

matriculas_cet.columns

cols = ['id_turma', 'discente', 'id_curso', 'media_final', 'numero_total_faltas', 'descricao']
matriculas_cet = matriculas_cet[cols]

## Removendo dados duplicados e vazios

matriculas_cet = matriculas_cet.drop_duplicates()
matriculas_cet.shape

matriculas_cet = matriculas_cet.dropna(subset=['id_turma', 'discente', 'descricao'])
matriculas_cet.shape

## Exportando os dados

matriculas_cet.to_csv("data/outputs/matriculas_cet_20162.csv", sep=";", index=False)
