==> 02 - Mesclando dados Matriculas, turmas e disciplinas.ipynb <==
# Mesclando os dados de Matrículas, Turmas e Disciplinas
## Objetivo
Esse notebook visa pegar informações que relacionam turmas, matrículas e disciplinas para facilitar 
a visualização dos dados.

import pandas as pd

matr = pd.read_csv("data/outputs/matriculas_cet_20162.csv", sep=';')
#disciplinas = pd.read_csv("data/componentes-curriculares-presenciais-ok.csv", index_col=0)
turmas = pd.read_csv("data/turmas-20162.csv", sep=";")

matr.head()

turmas.head(2)

## Pegando só as colunas desejadas

turmas.columns

cols = ['id_turma', 'codigo', 'ano', 'periodo']
turmas = turmas[cols]

## Realizando merge de Matriculas e Turmas

matr.shape

matriculas = matr.merge(turmas, how='left', left_on='id_turma', right_on='id_turma')
matriculas.shape

matriculas = matriculas.drop_duplicates()
matriculas.shape

# matr_ect.groupby('nome')['nome'].count().sort_values(ascending=False)

## Exportando os dados

matriculas.to_csv("data/outputs/matr_final_20162.csv", sep=";", index=False)
