==> 03 - Agrupando os estados.ipynb <==
# Agrupandos dos diferentes estados dos alunos de C&T
## objetivo
O objetivo desse notebook é lista todos os estados possíveis de um aluno em C&T, tendo como vista
as materias que ele pode pagar nesse curso. Serão listadas apenas as disciplinas oferecidas pelo departamento, 
tendo em vista que o objetivo é listar os primeiros semestres, onde ainda não há uma necessidade tão 
grande de pagar matérias em outros departamentos.

## Definição de Estado
Um Estado é formado pelo conjunto de disciplinas que o discente está pagando naquele semestre, assim, haverá um
estado para cada possibilidade possível de disciplinas. Os estados com menor adesão serão tidos como outlies e 
não serão utilizados na aplicação do algoritmo.

import pandas as pd
import numpy as np

matr = pd.read_csv("data/outputs/matr_final_20162.csv", sep=";")

matr.head()

## Agrupando por Discentes
Nesse passo, vamos agrupar as matriculas por discentes a fim de começar a mapear os estados e 
verificar possíveis outliers.

# matr.groupby('discente')['discente'].count().sort_values(ascending=False)

matr.descricao.unique()

matr.sort_values('codigo', ascending=True, inplace=True)

simple_desc = {'REPROVADO POR MÉDIA E POR FALTAS': 'R', 'REPROVADO POR NOTA': 'R',
                 'REPROVADO POR NOTA E FALTA': 'R',
                 'REPROVADO POR FALTAS': 'R', 'REPROVADO': 'R',
                'APROVADO POR NOTA': 'A', 'APROVADO': 'A', 'TRANCADO': 'T'}

matr['descricao'] = matr['descricao'].map(simple_desc)

group = matr[['discente', 'codigo', 'descricao']].groupby('discente')

group.agg(lambda x: ','.join(x))

## Criando os Estados
Nesse passo, vamos criar os estados dos discentes, sendo verificado as disciplinas que estão sendo pagas e 
perando um hash para cada conjunto de disciplinas.

Pode ser criado um discionários ou estrutura semelhante para armazenar os hashs que serão criados e agrupados.

hash(str(['ECT2414', 'ECT0123', 'ECT1241']))

def hash_situacao(descricao):
    '''
    Função para computar o resultado do discente no semestre.
    '''
    situacoes = list(descricao)
    situacoes.replace

def hash_disciplinas(disciplinas):
    '''
    Função que calcula um hash para as disciplinas. Assim será possível varer todos os estados possíveis.
    
    A função recebe uma lista de disciplinas e retorna um valor hash.
    >>> hash_disciplinas(['ECT2414', 'ECT0123', 'ECT1241'])
    '3-2914823815191669391'
    '''
    # O primeiro numero é a quantidade de disciplinas
    hash_ = str(len(disciplinas))
    
    # Ordena o vetor de disciplinas e gera uma string dele
    disc = list(disciplinas)
    disc.sort()
    hash_ += "".join(disc)
    
    return hash_

hash_disciplinas(['ECT2414', 'ECT0123', 'ECT1241'])

## Agrupando os discentes por Estado

# Aplicando a função hash_disciplinas no nosso grupo

estados = group.agg(hash_disciplinas)
type(estados)

estados.groupby('codigo')['codigo'].count().sort_values(ascending=False).head(5)

## Analisando só materias obrigatórias

obrig_s1 = ['ECT2101', 'ECT2102', 'ECT2103', 'ECT2104', 'ECT2105', 'ECT2106']
obrig_s2 = ['ECT2201', 'ECT2202', 'ECT2203', 'ECT2204', 'ECT2205', 'ECT2206', 'ECT2207']
obrig_s3 = ['ECT2301', 'ECT2302', 'ECT2303', 'ECT2304', 'ECT2305', 'ECT2306', 'ECT2307']
obrig_s4 = ['ECT2401', 'ECT2402', 'ECT2403', 'ECT2311', 'ECT2412', 'ECT2413', 'ECT2414']

obrig = obrig_s1 + obrig_s2 + obrig_s3 + obrig_s4

# Pegando apenas as matriculas em disciplinas obrigadorias
matr_obrig = matr[matr.codigo.isin(obrig)]
group_ob = matr_obrig[['discente', 'codigo', 'descricao']].groupby('discente')

estados_ob = group_ob.agg(lambda x: ','.join(x))

estados_ob.groupby('codigo')['codigo'].count().sort_values(ascending=False).head()

## Adicionando outros atributos

# Hash para disciplinas
estados_ob['hash'] = group_ob.codigo.agg(hash_disciplinas)

def periodo_ano(first):
    #first = df.iloc[0]
    return (str(first.ano) + str(first.periodo))

estados_ob['ano_periodo'] = periodo_ano(matr.iloc[0])

# Taxa de Aprovação
estados_ob['tx_aprovacao'] = group_ob.descricao.apply(lambda x: list(x).count('A')/len(list(x)))

# Exportando em csv

estados_ob.to_csv("data/outputs/estados_20162.csv", sep=";")

# estados_ob
