==> 04 - Merge dos semestres.ipynb <==
# Realizando o merge entre os semestres
## Objetivo
Esse notebook tem como fim fazer a junção entre vários semestre diferentes de C&T, começando em 2016.1

O motivo da escolha por 2016.1 é a mudança da grade curricular que entrou em vigor nesse semestre.

import pandas as pd
import numpy as np

df_matr_20161 = pd.read_csv("data/outputs/estados_20161.csv", sep=";")
df_matr_20162 = pd.read_csv("data/outputs/estados_20162.csv", sep=";")
df_matr_20171 = pd.read_csv("data/outputs/estados_20171.csv", sep=";")
df_matr_20172 = pd.read_csv("data/outputs/estados_20172.csv", sep=";")

## Tratando Estado

matr = df_matr_20172

estados_matr = pd.DataFrame()
estados_matr['hash'] = matr.groupby('hash').count()['discente'].index
estados_matr['qnt_discentes'] = matr.groupby('hash').count()['discente'].values

estados_matr['tx_aprovacao'] = matr.groupby('hash').mean()['tx_aprovacao'].values
estados_matr['periodo'] = matr.groupby('hash').mean()['ano_periodo'].values

## Realiando o Merge

# DataFrame Geral de Estados
# ------ só roda uma vez ----

estados_geral = pd.DataFrame(columns=estados_matr.columns)

estados_geral = pd.concat([estados_geral, estados_matr])

estados_geral.shape

estados_geral.to_csv("data/outputs/estados_processados/estados_final.csv", sep=";")
