==> 05 - Criando um modelo.ipynb <==
## Criando um modelo para os dados

O objetivo desse notebook é criar um modelo de dados para realizar predições com os dados já
capturados.

import pandas as pd
import numpy as np

data = pd.read_csv("data/outputs/estados_processados/estados_final.csv", sep=";")
data = data.iloc[:, 1:]

data.groupby('hash')['hash']

data.sort_values('qnt_discentes', ascending=False)
