==> 06 - Análise primeiro semestre.ipynb <==
# Análisando o Primeiro Semestre
Gerando resultados a partir dos dados dos discentes do Primeiro Bimestre

import pandas as pd
import numpy as np

matr_1 = pd.read_csv("data/outputs/estados_20161.csv", sep=";")
matr_2 = pd.read_csv("data/outputs/estados_20162.csv", sep=";")

semest_1 = matr_1[matr_1.hash == '6ECT2101ECT2102ECT2103ECT2104ECT2105ECT2106']
semest_1.shape

semest_2 = matr_2[matr_2.discente.isin(semest_1.discente)]
semest_2.shape

estado_atual = semest_2.groupby('hash').count()['codigo']
estado_atual.sort_values(ascending=False).head(10)

media_estado_atual = estado_atual / semest_2.shape[0]
media_estado_atual.sort_values(ascending=False).head(10)

# Taxa de Aprovação 1º Semestre

desempenho1 = semest_1[semest_1.hash == '6ECT2101ECT2102ECT2103ECT2104ECT2105ECT2106']
desempenho1.tx_aprovacao.mean()

ap_tudo = desempenho1[desempenho1.tx_aprovacao == 1.0].shape[0]
total = desempenho1.shape[0]
print(str(ap_tudo) + " Foram aprovados em tudo de " + str(total))

## Analisando Desempenho de quem passou em tudo

desempenho = semest_2[semest_2.hash == '7ECT2201ECT2202ECT2203ECT2204ECT2205ECT2206ECT2207']
desempenho.tx_aprovacao.mean()

ap_tudo = desempenho[desempenho.tx_aprovacao == 1.0].shape[0]
total = desempenho.shape[0]
print(str(ap_tudo) + " Foram aprovados em tudo de " + str(total))

print("Probabilidade de Sucesso em todas as disciplinas: "+str(ap_tudo/total))

## Próximos passos
- Análise de todos os estados
- Agrupamentos dos estados
- Utilização da *Cadeia de Markov*
